refactor: move Cognite3DModel
