+++ viewer/core/index.ts
@@ -2,6 +2,6 @@
  * Copyright 2021 Cognite AS
  */
 export { Cognite3DViewer } from './src/public/migration/Cognite3DViewer';
-export { Cognite3DModel } from './src/public/migration/Cognite3DModel';
+export { Cognite3DModel } from '@reveal/cad-model';
 
 export * from './src/public/types';
