+++ viewer/core/src/index.ts
@@ -11,10 +11,9 @@
  */
 
 export { revealEnv, IndexSet, NumericRange } from '@reveal/utilities';
-export { BoundingBoxClipper } from '@reveal/cad-model';
 
 export * from './public/migration/Cognite3DViewer';
-export { Cognite3DModel } from './public/migration/Cognite3DModel';
+export { Cognite3DModel, BoundingBoxClipper } from '@reveal/cad-model';
 export { Cognite3DViewer } from './public/migration/Cognite3DViewer';
 export { CognitePointCloudModel } from './public/migration/CognitePointCloudModel';
