+++ viewer/core/src/migration.ts
@@ -2,7 +2,7 @@
  * Copyright 2021 Cognite AS
  */
 
-export { Cognite3DModel } from './public/migration/Cognite3DModel';
+export { Cognite3DModel } from '@reveal/cad-model';
 export { CognitePointCloudModel } from './public/migration/CognitePointCloudModel';
 export { Cognite3DViewer } from './public/migration/Cognite3DViewer';
 export {
