+++ viewer/core/src/utilities/ViewStateHelper.ts
@@ -3,7 +3,7 @@
  */
 import * as THREE from 'three';
 
-import { Cognite3DModel } from '../public/migration/Cognite3DModel';
+import { Cognite3DModel } from '../../../packages/cad-model/src/Cognite3DModel';
 import { Cognite3DViewer } from '../public/migration/Cognite3DViewer';
 import { NodeCollectionDeserializer } from '@reveal/cad-styling';
