+++ viewer/packages/cad-model/index.ts
@@ -9,3 +9,4 @@ export { intersectCadNodes } from './src/picking';
 export { createCadManager } from './src/createCadManager';
 export { BoundingBoxClipper } from './src/BoundingBoxClipper';
 export { InternalRevealCadOptions } from './src/InternalRevealCadOptions';
+export { Cognite3DModel } from './src/Cognite3DModel';
