+++ viewer/packages/cad-model/src/WellKnownUnit.ts
@@ -1,0 +1,20 @@
+/*!
+ * Copyright 2022 Cognite AS
+ */
+
+
+/**
+ * Units supported by {@link Cognite3DModel}.
+ */
+export type WellKnownUnit =
+  | 'Meters'
+  | 'Centimeters'
+  | 'Millimeters'
+  | 'Micrometers'
+  | 'Kilometers'
+  | 'Feet'
+  | 'Inches'
+  | 'Yards'
+  | 'Miles'
+  | 'Mils'
+  | 'Microinches';
