+++ viewer/packages/model-base/index.ts
@@ -5,3 +5,5 @@
 export { IntersectInput } from './src/types';
 export { SupportedModelTypes } from './src/SupportedModelTypes';
 export { CogniteModelBase } from './src/CogniteModelBase';
+
+export { NodeIdAndTreeIndexMaps } from './src/NodeIdAndTreeIndexMaps';
